Add more validation to FlightRepository

+++ FlightAPI/Repositories/Implementations/FlightRepository.cs
@@ -5,16 +5,25 @@ using FlightAPI.Models;
 using FlightAPI.Models.DTOs;
 using FlightAPI.Repositories.Interfaces;
 using Microsoft.EntityFrameworkCore;
+using FlightAPI.Exceptions;
+using FlightAPI.Services.Implementations;
 
 namespace FlightAPI.Repositories.Implementations
 {
-    public class FlightRepository(IApplicationDbContext db, IMapper mapper) : IFlightRepository
+    public class FlightRepository(IApplicationDbContext db, IMapper mapper, ILogger<FlightRepository> logger) : IFlightRepository
     {
         private readonly IApplicationDbContext _db = db;
         private readonly IMapper _mapper = mapper;
+        private readonly ILogger<FlightRepository> _logger = logger;
 
         public async Task<FlightDTO> Create(CreateFlightDTO flightDTO)
         {
+            if(flightDTO is null)
+            {
+                _logger.LogError("Attempted to create a flight with null data");
+                throw new NullFlightDataException();
+            }
+
             Flight newFlight = new()
             {
                 FlightNumber = flightDTO.FlightNumber,
@@ -32,6 +41,12 @@ namespace FlightAPI.Repositories.Implementations
 
         public async Task Delete(Flight flight)
         {
+            if (flight is null)
+            {
+                _logger.LogError("Attempted to delete a flight with null data");
+                throw new NullFlightDataException();
+            }
+
             _db.Flights.Remove(flight);
             await _db.SaveChangesAsync();
         }
@@ -46,7 +61,7 @@ namespace FlightAPI.Repositories.Implementations
 
         public async Task<Flight?> GetFlightById(int id)
         {
-            return await _db.Flights.FindAsync(id);
+            return await _db.Flights.FirstOrDefaultAsync(f => f.Id == id);
         }
 
         public async Task<FlightDTO?> GetFlightDTOById(int id)
@@ -59,10 +74,20 @@ namespace FlightAPI.Repositories.Implementations
 
         public async Task<FlightDTO> Update(UpdateFlightDTO flightDTO, Flight flight)
         {
+            if(flight is null || flightDTO is null)
+            {
+                _logger.LogError("Attempted to update a flight with null data");
+                throw new NullFlightDataException();
+            }
+
             _mapper.Map(flightDTO, flight);
             _db.Flights.Update(flight);
             await _db.SaveChangesAsync();
-            flight.Plane = await _db.Planes.FindAsync(flight.PlaneId);
+
+            if (flight.PlaneId != null)
+            {
+                flight.Plane = await _db.Planes.FirstOrDefaultAsync(p => p.Id == flight.PlaneId);
+            }
 
             return _mapper.Map<FlightDTO>(flight);
         }
